DEV-12701 Further refinement to better separate logic and adding tests etc
